+++ audiotools/filter/butterworth_filt.py
@@ -2,12 +2,14 @@ import scipy.signal as sig
 import numpy as np
 from .. import audiotools as audio
 
-def _copy_to_dim(array, dim):
-    if np.ndim(dim) == 0: dim = (dim,)
 
-    #tile by the number of dimensions
+def _copy_to_dim(array, dim):
+    if np.ndim(dim) == 0:
+        dim = (dim,)
+
+    # tile by the number of dimensions
     tiled_array = np.tile(array, (*dim[::-1], 1)).T
-    #squeeze to remove axis of lenght 1
+    # squeeze to remove axis of lenght 1
     tiled_array = tiled_array
 
     return tiled_array